pyproject.toml
[project]
name = "plantgenie-api"
version = "0.1.0"
description = "Backend for the PlantGenIE Web Application (and general usage)"
readme = "README.md"
authors = [{ name = "Jamie McCann", email = "jamie.mccann@umu.se" }]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "celery[pydantic,redis]>=5.5.0",
    "duckdb>=1.2.0",
    "fastapi[standard]>=0.115.11",
    "loguru>=0.7.3",
    "pendulum>=3.1.0",
    "py-fasta-validator>=0.6",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "python-keystoneclient>=5.6.0",
    "python-multipart>=0.0.20",
    "python-swiftclient>=4.8.0",
    "redis>=5.2.1",
    "python-dotenv>=1.0.1",
    "aiosqlite>=0.22.1",
    "shared",
    "task-queue",
    "go-enrich"
]

[tool.uv.workspace]
members = [
    "packages/task-queue",
    "packages/shared",
    "packages/go-enrich",
]

[tool.uv.sources]
task-queue = { workspace = true }
shared = { workspace = true }
go-enrich = { workspace = true }

[project.scripts]
plantgenie-api = "plantgenie_api.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "anyio>=4.9.0",
    "black>=25.1.0",
    "celery-types>=0.23.0",
    "ipython>=9.3.0",
    "mypy>=1.15.0",
    "pytest>=9.0.1",
    "python-dotenv[cli]>=1.0.1",
    "ruff>=0.9.5",
]

[tool.black]
line-length = 75

[tool.pytest.ini_options]
minversion = "9.0"
testpaths = [
    "src/tests",
    "packages/task-queue/src/tests",
    "packages/shared/src/tests",
]
addopts = ["--import-mode=importlib"]
filterwarnings = ["ignore::DeprecationWarning:testcontainers"]
